chore: Refactor functions.
